+++ pdf_extract.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 """
 pdf_extract.py
 
@@ -12,7 +12,7 @@ from PyPDF2 import PdfFileReader, PdfFileWriter
 import sys
 import shlex
 import os.path
-import re 
+import re
 
 if __name__ == "__main__":
     pdfs = []
@@ -30,18 +30,43 @@ if __name__ == "__main__":
     nb_pages = pdf.numPages
     print(f"Nombre de pages trouvées : {nb_pages}")
 
-
-
-    pages_txt = input("Numéros des pages à extraire (séparés par un espace) : ")
+    pages_txt = input("Pages à extraire : ")
     pages = shlex.split(pages_txt)
+    if pages_txt.strip() == "*":
+        # On souhaite extraire toutes les pages.
+        pages = [str(x) for x in range(1, nb_pages + 1)]
+
     for p in pages:
-        if not p.isnumeric() or int(p) > nb_pages:
-            print(f"\"{p}\" n'est pas une page valide.")
-            tout_est_ok = False
+        subpdf = re.match(r"(\d+)-(\d)", p)
+        tout_est_ok = True
+        if subpdf:
+            # On souhaite extraire des pages continues.
+            subpdf_from = subpdf[1]
+            subpdf_to = subpdf[2]
+
+            if (
+                not subpdf_from.isnumeric()
+                or int(subpdf_from) > nb_pages
+                or not subpdf_to.isnumeric()
+                or int(subpdf_to) > nb_pages
+            ):
+                print(f'"{p}" n\'est pas une plage de pages valide.')
+                tout_est_ok = False
         else:
+            # On souhaite extraire une page seule.
+            if not p.isnumeric() or int(p) > nb_pages:
+                print(f'"{p}" n\'est pas une page valide.')
+                tout_est_ok = False
+
+        if tout_est_ok == True:
             pdfWriter = PdfFileWriter()
-            pdfWriter.addPage(pdf.getPage(int(p) - 1))
-            file_out = re.sub('.pdf$', f"_{p}.pdf", file_in, flags=re.IGNORECASE)
+            if subpdf:
+                for sub_page in range(int(subpdf_from), int(subpdf_to) + 1):
+                    pdfWriter.addPage(pdf.getPage(sub_page - 1))
+            else:
+                pdfWriter.addPage(pdf.getPage(int(p) - 1))
+
+            file_out = re.sub(".pdf$", f"_{p}.pdf", file_in, flags=re.IGNORECASE)
 
             answer = ""
             while os.path.isfile(file_out) and answer not in ["O", "Y", "N"]:
@@ -50,11 +75,11 @@ if __name__ == "__main__":
                 ).upper()
 
             if answer.upper() in ["", "O", "Y"]:
-                with open(file_out, 'wb') as f:
-                    print(f"Extraction de \"{file_out}\" ", end="")
+                with open(file_out, "wb") as f:
+                    print(f'Extraction de "{file_out}" ', end="")
                     pdfWriter.write(f)
                     f.close()
                     print("OK")
-                
+
     print("Terminé !")
     os.system("pause")